go spring boot

+++ src/test/java/com/mzherdev/salesavingsystem/mock/ProductTestData.java
@@ -1,9 +1,10 @@
 package com.mzherdev.salesavingsystem.mock;
 
-import com.mzherdev.salesavingsystem.model.Product;
-
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
+
+import com.mzherdev.salesavingsystem.model.Product;
 
 public class ProductTestData {
 
@@ -24,16 +25,16 @@ public class ProductTestData {
 
     public static final int UNKNOWN_PRODUCT_ID = 100;
 
-    public static final Product JEANS = new Product(JEANS_PRODUCT_ID, "Jeans", 9.5);
-    public static final Product T_SHIRT = new Product(T_SHIRT_PRODUCT_ID, "T-shirt", 5.0);
-    public static final Product PULLOVER = new Product(PULLOVER_PRODUCT_ID, "Pullover", 39.99);
-    public static final Product SCARF = new Product(SCARF_PRODUCT_ID, "Scarf", 2.4);
-    public static final Product SHORTS = new Product(SHORTS_PRODUCT_ID, "Shorts", 2.5);
-    public static final Product SUNGLASSES = new Product(SUNGLASSES_PRODUCT_ID, "Sunglasses", 10.0);
-    public static final Product SHOES = new Product(SHOES_PRODUCT_ID, "Shoes", 25.0);
-    public static final Product CUFFLINKS = new Product(CUFFLINKS_PRODUCT_ID, "Cufflinks", 1.65);
-    public static final Product COAT = new Product(COAT_PRODUCT_ID, "Coat", 52.0);
-    public static final Product BOOTS = new Product(COAT_PRODUCT_ID, "Boots", 30.5);
+    public static final Product JEANS = new Product(JEANS_PRODUCT_ID, "Jeans", BigDecimal.valueOf(9.5));
+    public static final Product T_SHIRT = new Product(T_SHIRT_PRODUCT_ID, "T-shirt", BigDecimal.valueOf(5.0));
+    public static final Product PULLOVER = new Product(PULLOVER_PRODUCT_ID, "Pullover", BigDecimal.valueOf(39.99));
+    public static final Product SCARF = new Product(SCARF_PRODUCT_ID, "Scarf", BigDecimal.valueOf(2.4));
+    public static final Product SHORTS = new Product(SHORTS_PRODUCT_ID, "Shorts", BigDecimal.valueOf(2.5));
+    public static final Product SUNGLASSES = new Product(SUNGLASSES_PRODUCT_ID, "Sunglasses", BigDecimal.valueOf(10.0));
+    public static final Product SHOES = new Product(SHOES_PRODUCT_ID, "Shoes", BigDecimal.valueOf(25.0));
+    public static final Product CUFFLINKS = new Product(CUFFLINKS_PRODUCT_ID, "Cufflinks", BigDecimal.valueOf(1.65));
+    public static final Product COAT = new Product(COAT_PRODUCT_ID, "Coat", BigDecimal.valueOf(52.0));
+    public static final Product BOOTS = new Product(COAT_PRODUCT_ID, "Boots", BigDecimal.valueOf(30.5));
 
     public static final List<Product> ALL = Arrays.asList(JEANS, T_SHIRT, PULLOVER,
             SCARF, SHORTS, SUNGLASSES, SHOES, CUFFLINKS, COAT, BOOTS);
